misc: Use the code format for commands and messages
Make the style consistent across code blocks.  This also makes sure
that markdown does not try to read any of the code blocks and try to
do anything with it.

Using the blockquotes instead of code blocks results in github
attempting to render tags and special markdown chars, which is wrong.

Change-Id: I8526933c8283ae02646d6273ed0cbd8e85176bd0
Signed-off-by: Siddhesh Poyarekar <siddhesh.poyarekar@gmail.com>
Reviewed-on: https://gem5-review.googlesource.com/8323
Reviewed-by: Brandon Potter <Brandon.Potter@amd.com>
Reviewed-by: Jason Lowe-Power <jason@lowepower.com>
Maintainer: Jason Lowe-Power <jason@lowepower.com>