scons: remove as version check
GNU as 2.31.1 checks if the input and output files are the same:

Assembler messages:
Fatal error: The input and output files must be distinct

The check already does not work for llvm-as which has a different
output format.

Since it is too hard to maintain the check correctly for all possible
assemblers, it is better to just remove it completely.

Change-Id: I38a993ab83ca83d4a2f5e77820d2ca903f70c6ac
Reviewed-on: https://gem5-review.googlesource.com/12403
Reviewed-by: Jason Lowe-Power <jason@lowepower.com>
Reviewed-by: Andreas Sandberg <andreas.sandberg@arm.com>
Maintainer: Andreas Sandberg <andreas.sandberg@arm.com>