arch-arm: Fixed ARM/gem5.fast compilation failures
The compiler-tests were failing:
https://jenkins.gem5.org/job/compiler-checks/238

This was due to an `error: unused variable` error being thrown in cases
where a variable was declared and used soley in an `assert` within a
SLICC file. Assertions of this kind are stripped during .fast
compilation.

This patch fixes this.

Change-Id: I3a91ac8b1a51de7ddffd6a1cff602a934862b49c
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/59829
Maintainer: Bobby Bruce <bbruce@ucdavis.edu>
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: kokoro <noreply+kokoro@google.com>