misc: Split up tests in daily-tests.yaml
This splits up the gem5 library example tests by Suite UID, as
right now running them together uses the runner for a long
period of time.  It is important to note that doing this means
additional tests from this directory will need to be
manually added, such as the kvm tests.

Change-Id: Ib2a0aca08f9b51b60e9dd0528324372cf2d98c05